refactor(worker): move execute_command to Cli impl and extract signal handling
