+++ crates/shared/src/utils/mod.rs
@@ -4,6 +4,7 @@ use async_trait::async_trait;
 use std::sync::Arc;
 use tokio::sync::Mutex;
 pub mod google_cloud;
+pub mod signal;
 use anyhow::Result;
 
 #[async_trait]
