--- crates/worker/src/cli/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod command;
-pub use command::{execute_command, Cli};
